requirements.txt
# Python < 3.13 required (some ursa dependencies do not support 3.13+)
ursa-ai==0.15.1
langchain>=0.3
langchain-core
python-dotenv>=1.0
numpy
matplotlib
h5py
